Se movieron los listeners y observers a Controller
