+++ src/main/java/org/itson/mvc/Pool/PoolController.java
@@ -6,9 +6,13 @@ package org.itson.mvc.Pool;
 
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
+import java.util.ArrayList;
+import java.util.List;
 import javax.swing.SwingUtilities;
 import org.itson.domaincomponent.domain.Tile;
 import org.itson.game.MatchGame;
+import org.itson.interfaces.Observer;
+
 
 /**
  *
@@ -17,15 +21,16 @@ import org.itson.game.MatchGame;
 public class PoolController  extends MouseAdapter {
     private PoolView poolView;
     private PoolModel poolModel;
+    private List<Observer> observers = new ArrayList<>();
 
     public PoolController(PoolView poolView, PoolModel poolModel) {
         this.poolView = poolView;
         this.poolModel = poolModel;
-
+        this.suscribeToClick();
     }
     
-    public void suscribeToView(MatchGame match){
-        this.poolView.addObserver(match);
+    public void suscribe(MatchGame match){
+        this.addObserver(match);
     }
     
     public void createDominoTiles(){
@@ -40,5 +45,28 @@ public class PoolController  extends MouseAdapter {
         this.poolView.refresh();
     }
     
+    private void suscribeToClick(){
+        this.poolView.addMouseListener(new java.awt.event.MouseAdapter() {
+                @Override
+                public void mouseClicked(java.awt.event.MouseEvent evt) {
+                    notifyObservers("click_event");
+                }
+            });
+    }
 
+    
+    
+     public void addObserver(Observer observer) {
+        observers.add(observer);
+    }
+
+    public void removeObserver(Observer observer) {
+        observers.remove(observer);
+    }
+
+    public void notifyObservers(String message) {
+        for (Observer observer : observers) {
+            observer.update(message);
+        }
+    }
 }
