+++ src/main/java/org/itson/mvc/Pool/PoolView.java
@@ -18,7 +18,7 @@ public class PoolView extends JPanel {
     private Image poolImage;
     private int poolWidth;
     private int poolHeight;
-    private List<Observer> observers = new ArrayList<>();
+    
 
     public PoolView(PoolModel poolModel) {
         this.poolModel = poolModel;
@@ -30,13 +30,6 @@ public class PoolView extends JPanel {
             poolImage = ImageIO.read(new File(poolModel.getPoolImagePath()));
             poolWidth = 120;
             poolHeight = 130;
-
-            this.addMouseListener(new java.awt.event.MouseAdapter() {
-                @Override
-                public void mouseClicked(java.awt.event.MouseEvent evt) {
-                    notifyObservers("click_event");
-                }
-            });
 
         } catch (IOException e) {
             e.printStackTrace();
@@ -58,18 +51,6 @@ public class PoolView extends JPanel {
         g2d.drawImage(poolImage, 0, 0, poolWidth, poolHeight, this);
     }
 
-    public void addObserver(Observer observer) {
-        observers.add(observer);
-    }
-
-    public void removeObserver(Observer observer) {
-        observers.remove(observer);
-    }
-
-    public void notifyObservers(String message) {
-        for (Observer observer : observers) {
-            observer.update(message);
-        }
-    }
+   
 
 }
